Some random website changes (#2319)
Some random website changes (#2319)

* Remove blog

We're not going to have any content here at launch.

* Remove Community and GitHub

We're not going to advertise the sorbet-ruby Slack, and we're not going
to be open source at launch.

* Add "Get started" link to the header

The header was looking pretty bland:

    Sorbet                                                   Docs   Try

Now it looks a little bit more engaging:

    Sorbet                           Get started   Docs   Try it online

(Squashed by Merge Bot)