Changes LSP message params/results to a discriminated union rather than an any type. (#273)
**Overview**

Adds the ability to define the type of an object field as a specific JSON shape depending on the value of another object field (the discriminant).

The field receives the C++ type variant. At JSON parsing/serialization time, the generated code checks that the variant type holds the correct type given the current value of the discriminant.

The discriminant for NotificationMessage and RequestMessage is the method field. ResponseMessages lack a discriminant, but the type of result depends on the request, so we lurk a requestMethod field on the objects. Clients appear to ignore this extra field.

**What this gets us**

With these changes, Sorbet now checks the following properties at parsing/serialization time:

Request messages and notification messages use the correct subset of LSPMethods (e.g., so you cannot send/receive a DidChange as a request).
Request messages, notification messages, and response messages contain the correct shape of params/result.
These checks actually found a few bugs in tests!

**Downsides**

When two methods have a result that can take on multiple shapes (with one overlapping), the assignment to responseMsg.result needs to be explicitly typed to ensure that the assignment goes into the correct variant 'slot'.

For example, textDocument/Hover can be null | Hover, textDocument/references can be null | Location[], and shutdown is null. Assigning null to a response fills the shutdown slot, failing serialization errors at runtime for the other two.

The solution is to assign it with an explicit type:

```
response.result = variant<JSONNullObject, vector<unique_ptr<Location>>>(JSONNullObject());
```

Fixing this problem so that a regular null assignment "does the right thing" is somewhat challenging. If all result types were variants or single types, it would be easy to "inline" all of the possible variants. Unfortunately, some discriminated unions contain optional types -- so the same problem would occur with null | optional<null>.

Another solution is to define entirely new classes for all possible Request/Response/Notification shapes (e.g., TextDocumentDidOpenRequest). This is possible, but makes dealing with messages in the abstract a bit more challenging. The LSPMessage convenience class would have to wrap and know about all possible shapes, and have if / elses to cover them.

I think that my design is a decent trade off. If we ensure that our LSP features have sufficient test coverage, we can ensure that this doesn't become a problem in practice.

**What this enables**

This is the biggest change needed to delurk our use of the JSONAny type and passing around rapidjson objects and allocators. A subsequent PR will do just that, fixing our largest memory leak and forcing all deserialization errors to happen "up front" (rather than middle-of-execution, when LSP deserializes an any type).

The discriminated union will also make it possible to merge Watchman and editor file updates easily and without leaking memory.