Remove error when running textDocument/definition on untyped file (#258)
* Remove error when running textDocument/definition on untyped file.

The UX for this error is really bad. The extension output steals focus, and VS Code displays a message. VS Code makes these requests on behalf of the user in the background, so they can happen at any moment.

Since go-to-definition is disabled unless the user explicitly opts into it, I believe returning the empty set is acceptable behavior.