Make LSP test_corpus tests run a second time through the fast path. (#322)
* Make LSP test_corpus tests run a second time through the fast path.

Ensures that output from the fast path agrees with the slow path. Assertions that fail on the fast path are clearly labeled as such.

To prevent duplicate/spammy test failures, the fast path isn't run if the slow path fails.

Includes an opt-out for current (and future) tests that fail this check.