Add flag to ignore config file
The only way around this right now is to cd to a different directory or
to move the config file out of the way. I think we should have an option
to opt out of this if people want it. It's also something we can suggest
people do when reproducing issues.