Make toString, toStringFull, showRaw, and showRawFull
For example:

- ast, ast-raw
- dsl-tree, dsl-tree-raw
- name-tree, name-tree-raw
- resolve-tree, resolve-tree-raw

These options work with the convention that "raw" debugging output
is not often the prettiest / concisest representation. We use the names
`toString` and `showRaw` to handle this distinction right now.

The symbol table printing doesn't conform to that convention, so this PR
fixes it, across a number of changes.

In this first change, we fix GlobalState to have `toString` and
`showRaw` methods. GlobalState is somewhat special in that it also has a
second option: whether to show or hide certain symbols. We've chosen the
options `symbol-table-full` and `symbol-table-raw-full` to handle these
cases, so I've also created names for these methods. In total:

- `toString`
- `toStringFull`
- `showRaw`
- `showRawFull`

Also, this has the benefit that it's easier to remember, esp. in the
REPL ("Which boolean came first? What's the default?").