Propagate showRaw / toString changes to Symbol
Symbol had the same problems that GlobalState did: there was no showRaw,
only a toString and a toStringWithTabs (that did more than just tabs).

Symbol is even more special, because we sometimes need to be able to
show a Symbol to a user. So a Symbol has three methods:

- `showRaw`
- `toString`
- `show`

These three methods are as follows:

`show` is the familiar "something safe to show to the user".

`toString` means "we were called from -p symbol-table" or otherwise
"please abbreviate the output"

`showRaw` means "we were asked to be as verbose as possible, exposing
the raw structure."

I've also taken the opportunity to make toStringWithOptions private,
because no one needed the full generality of before, except GlobalState,
which is a friend of Symbol.