Name::toString -> Name::showRaw
The next smell is that when we're printing Symbols / the symbol-table:

- `symbol-table-raw` means use `toString`
- `symbol-table` means use `show`

For all other paired data structures, it looks like this:

- `symbol-table-raw` means `showRaw`
- `symbol-table` means `toString`

While we could probably come up with a slightly better `toString` than
`show`, using `show` for the non-raw output is actually ~ok right now.

So to get back up to parity with the rest of the codebase, we'll rename
`toString` -> `showRaw`. This makes sense, because this method is really
gnarly... it's the one that actually prints `<C <U <foo>>`, which is
about as raw as a Name can get.