Refactor toString and showRaw for consistency (#418)
* Make toString, toStringFull, showRaw, and showRawFull

For example:

- ast, ast-raw
- dsl-tree, dsl-tree-raw
- name-tree, name-tree-raw
- resolve-tree, resolve-tree-raw

These options work with the convention that "raw" debugging output
is not often the prettiest / concisest representation. We use the names
`toString` and `showRaw` to handle this distinction right now.

The symbol table printing doesn't conform to that convention, so this PR
fixes it, across a number of changes.

In this first change, we fix GlobalState to have `toString` and
`showRaw` methods. GlobalState is somewhat special in that it also has a
second option: whether to show or hide certain symbols. We've chosen the
options `symbol-table-full` and `symbol-table-raw-full` to handle these
cases, so I've also created names for these methods. In total:

- `toString`
- `toStringFull`
- `showRaw`
- `showRawFull`

Also, this has the benefit that it's easier to remember, esp. in the
REPL ("Which boolean came first? What's the default?").

* Propagate showRaw / toString changes to Symbol

Symbol had the same problems that GlobalState did: there was no showRaw,
only a toString and a toStringWithTabs (that did more than just tabs).

Symbol is even more special, because we sometimes need to be able to
show a Symbol to a user. So a Symbol has three methods:

- `showRaw`
- `toString`
- `show`

These three methods are as follows:

`show` is the familiar "something safe to show to the user".

`toString` means "we were called from -p symbol-table" or otherwise
"please abbreviate the output"

`showRaw` means "we were asked to be as verbose as possible, exposing
the raw structure."

I've also taken the opportunity to make toStringWithOptions private,
because no one needed the full generality of before, except GlobalState,
which is a friend of Symbol.

* Name::toString -> Name::showRaw

The next smell is that when we're printing Symbols / the symbol-table:

- `symbol-table-raw` means use `toString`
- `symbol-table` means use `show`

For all other paired data structures, it looks like this:

- `symbol-table-raw` means `showRaw`
- `symbol-table` means `toString`

While we could probably come up with a slightly better `toString` than
`show`, using `show` for the non-raw output is actually ~ok right now.

So to get back up to parity with the rest of the codebase, we'll rename
`toString` -> `showRaw`. This makes sense, because this method is really
gnarly... it's the one that actually prints `<C <U <foo>>`, which is
about as raw as a Name can get.

* toString -> showRaw (all call sites)

* Add back Name::toString

This restores Name::toString to essentially as it existed before
58b220e6509a4 changed the things to use the `<U ...>` syntax.

This syntax was

- less raw
- still raw enough to have more information about which things were
  constant names vs utf8 names

It also adds a forwarder on LocalVariable (which just wraps a Name).

* fixup typo

Co-Authored-By: jez <zimmerman.jake@gmail.com>