Make all forks absolute
We don't want to rely neither on things being on the path nor on gems
being installed, because this makes it harder to test locally.

Key observation: once any piece of our code is running (either `srb` or
`srb-rbi`), we can always compute the path to any other component we
need. This means we don't have to rely on rvm, rbenv, the PATH variable,
etc.

For most things, this is just a matter of sprinkling a bunch of `../../`
The one tricky thing is that if we can't find the path to `sorbet` in
`sorbet-static`, we fall back to trying to find the version of Sorbet
that bazel has built in `bazel-bin/main/sorbet`.