Correct sig for Kernel#Array to make it usable
The old signature was broken and returns a T::Array[T.noreturn]
which makes methods that work on the elements in the array unusable.

The overloads I have here isn't ideal because if you pass
an T::Array[Thing] into it Sorbet returns an untyped array back.

If I add another overload to try to preserve the Array type
```ruby
  sig do
    type_parameters(:T).params(
        x: T::Array[T.type_parameter(:T)],
    )
    .returns(T::Array[T.type_parameter(:T)])
  end
```

Overload resolution isn't smart enough to be able to fallback to the
BasicObject overload when we pass it a value that isn't an array
(or nil).

The story is similar for hashes.

Please let me know if you have ideas for making this better.

Fixes #445