Pause tracepoints when requiring
trevor reported that he couldn't regenerate the snapshots on his laptop
because the tracepoint categorizer was bucketing the `bundler`
definitions into `site_ruby.rbi`. I'd rather not have a growing black
list of files to not serialize, especially when we know the root cause.

Unfortunately, this still doesn't let us serialize the bundler.rbi file.
I was hoping that if the project we're loading wanted to use bundler,
we'd hit the `gem 'bundler'` line in their `Gemfile`, and require it.

This does happen, but since we've already loaded bundler once, it
doesn't load a second time, so the tracepoint hooks don't see it. We
can't even get around by pinning a different version of bundler in the
`Gemfile.lock` to force it to load. Instead, bundler just refuses to
load the project.