Make pipeline.yaml names match filesystem
I've had more trouble than I'd like matching the names I see in the
Buildkite UI with the files that I have to edit to fix something.

I also would like to be able to see which jobs are running on which OS's
because that determines why my job is waiting.

I chose to make the names be the name of the file that's running that
part of the pipeline. This means the names will be somewhat less human
readable. But I think having the names be the files makes it easier for
new contributors to figure out which file they can look at to see why
their builds are failing.