Fix master breakage
We added a new RBI to sorbet-typed and broke master.
This is a known issue. The tracking ticket is #599.

Also, this seems like it's exposed a bug in sorbet: the Gemfile for this
test is completely empty. It does not mention 'bundler' at all, yet we
download an RBI for it.

I'll track down the cause of that later, but for now I think we should
unbreak master.