RFC: Better rvLoc using backEdges
Motivation

If the last statement a return came from an `if`, we previously weren't
smart enough to figure out the Loc for the binding that holds the return
value (because the basic block would be empty of statements, but have
multiple in-edges). In such cases, we fell back to using the Loc of the
entire method (which might be many, many lines long).

This manifested as error messages in LSP that turned the entire body of
the method red. And since LSP is fast now, it feels really intrusive to
have the entire WIP method body turn red.

Summary

This change adds a heuristic based approach to attempt to find a better
loc. Specifically, when there are no exprs in the last block, but it has
back edges (i.e., there are other blocks with forward edges here) we
pick one of these edges and use it.

This is a bad heuristic. If there are multiple blocks that could jump to
the return val, we should show them all when there's a type error in the
method return. I haven't though through how to do this properly, which
means we could still get bad error message locations.

**This is an RFC.** If we like this approach and want to pursue it
further, I'm happy digging in a little more, or helping you dig in a
little more.

An open question: can we think of an example where there are more than
two back-edges in the return value block? Specifically, we compute this
block **before simplification**, so it's not the case that we might
collapse multiple empty blocks to point at this block. This would have
to be specifically from the raw result of builder_walk.

Test plan

I've included some testdata tests that were useful to me when
prototyping this.

Do we have a way to write testdata tests that **assert multi-line
errors**? If so, I'd love to learn how.