master -> origin/master
We don't pull anymore, which means we don't merge `origin/master` into
`master`. This would have been a silent bug where the next time we had
updated `sorbet-typed`, we would have fetched an update, but not checked
it out, and then someone would have complained to us about not being
able to see the updates they pushed.

Test plan

I can't think of a way to add a regression test that would have
exercised the bad behavior except doing something like writing a script
that goes into the sorbet-typed cache folder and mucks with it to make
sure that `fetch-rbis` properly unmucks it after this change.

Open to suggestions.