Runtime: Make sig work in class << self
Previously, in order to use `sig` in `class << self`, one had to
call `T::Hooks.install(self)` before `class << self`. Failing to
do so can result in confusing errors like

> You called sig twice without declaring a method inbetween

This commit eliminate the need for calling `T::Hooks.install`
making `extend T::Sig` the only entry requirement.

Notes on implementation:

`install_hook` is called with the receiver of `sig` and for
`sig` to work, we need to hook the method definition that immediately
follows. Sorbet accepts both `def foo` and `def self.foo` and we to
handle both cases.

The following shows what hooks Ruby call:
```
class A
  class << self
    def foo; end # calls A.singleton_method_added
    def self.foo; end # calls A.singleton_class.singleton_method_added
  end
end
```

We catch `A.singleton_method_added` by defining it on
`A.singleton_class` and catch `A.singleton_class.singleton_method_added`
by installing it on `A.singleton_class.singleton_class`.

The code for handling when sig is not a singleton class should be
effectively untouched.