Quality of life changes for `srb init` (#786)
* Use '├─' instead of '└─' when we keep going

* warn -> attn

We're a bit allergic to "warnings" on this team. The test suite doesn't
actually ever mark a test as "succeeded with warnings" (only ever
succeeded or failed). This logging helper is more aptly described as
`attn` (attention), because it makes the whole line yellow.

* Make "Not recording empty partial test" quieter

It used to be two lines and one of them was all yellow, and would be
printed for ~every test when using `driver.sh --update`.

This reduces it to one line, and makes it just `info`, not `attn`.