Add extra note regarding `Boolean` for code 7001
The type change of a boolean variable is counter-intuitive for most people, thus, making the 7001 error harder to comprehend in those cases.

This PR adds a note about that specific case, that details how the type might change from `TrueClass` to `FalseClass` (or vice-versa) to make the 7001 error code a little more intelligible for those exact situations.