Add explicit --dir and --file options (#850)
This PR adds --dir and --file options to Sorbet, while retaining the existing CLI support for a list of paths as input. The benefit to these options is that they do not require a stat syscall for every file/directory input to Sorbet.

The PR also switches srb init to use --dir in the Sorbet config.