Make Hover non-experimental.
Hover works well enough to be enabled for everyone. If there is a bug, it is easy to add a new test case to Sorbet and iterate on a solution.

In real-world uses, hover does not appear to cause any issues with typechecking speed. In particular, VS Code cancels hover if you move your mouse away, so mousing over many things during a slow path won't interrupt edit batching.