[prework] Add types to struct initialize (#1024)
* Rm unused include

* Avoid `if not cond then ... else ...`

* Fix typo

* Have ChalkODMProp::replaceDSL return a optional

It is ENFORCEd that if ret == SOME(x) then x != [].

This is in preparation for allowing ChalkODMProp::replaceDSL to return
something other than the vector<unique_ptr<ast::Expression>>.
Previously, because the only thing we were returning was that vector, we
could use the empty vector to signify "no desugaring occurred".

* Use ChalkODMProp::{, NodesAnd}Prop

* Collect the props

* Wrap a comment

* Use T

* Update a comment

The old comment seemed to be inaccurate in multiple places. This
attempts to make it more accurate.

Of course, it's likely that the old comment was probably accurate at one
point and became inaccurate over time as the code changed. The same
could happen to this new comment. And since the inaccuracy of the old
comment wasn't noticed until now, it's arguable that we could remove the
comment altogether and not lose much.

* Rm ws

* Make *.rbi have *.rb syntax in vscode

* Use arg0

* Rm an arg

* Rm a comment invariant