Unbreak master
Because we didn't have a `Gemfile`, `rbenv exec srb` just runs whichever
version of `srb` is installed for the current Ruby version in
`.ruby-version`.

In particular, because the `~/.rbenv` folder is shared across tests, it
might have artifacts from other people's branches (specifically when the
tests failed, and we didn't get to the line to delete the existing
gems). And because our version numbers are just "how many commits since
the beginning of time", 0.4.FOO might be the same number as 0.4.BAR in
someone else's branch, even though those two versions are completely
unrelated, and **also** completely unrelated to the gems that will
EVENTUALLY be published with that as the Canonical Version.

I chose to both remove before each run and also trap EXIT just in case,
because in no cases do we want these gems to persist on the mac machine
after a CI run.