Add `require 'delegate'` to the tracepoint_serializer module
`require 'delegate'` happens automatically with full ruby builds, but
the version that we're building in bazel lacks `did_you_mean`, which
requires the delegate library. Requiring 'delegate' here will be a no-op
in most cases, but will allow the snapshot testing in bazel to proceed
without digging into the ruby build further.