Copy _was_modified when deep cloning
This flag is only used for caching purposes as of current.
The index result cache was not retained if this flag was
not set.

When plugins are enabled, we copy the global state more times than
usual. Since this flag was unset on every copy, it can happen that
we do not store the cache even though we would've if only plugins
are disabled.

I think this new semantic makes sense for the purposes of caching.