Output caching for DSL plugins
DSL plugins are extremely slow compare to other parts of Sorbet. This
commit mitigates the problem by caching the output of DSL plugins.

Before this commit, plugin outputs are only processed on cold runs. On
cached runs, no plugin outputs were considered. The output of plugins
and in term, their processed trees were not cached at all. The plugin
system generate separate virtual files which the caching system
discarded. Type checking results on cold and cache runs can be different
because of this limitation.

This commit caches the output of plugins, keyed on the hash of the file
that made the plugin call.  In addition, supplied files now cache the
number of plugin invocations they make. This is needed to compute the
cache key for each plugin output.

With these two new pieces in the cache, we avoid spawning subprocesses
when the cache is hot.

There is no cache invalidation based on the implementation of plugins at
all, as it's arbitrary Ruby code that could wild things, including
fetching things from the internet. In fact, there is no invalidation
based on the plugin spec yaml, either. These limitations make this
system unsuitable for use during the development of plugins.