Make NameRef remember the globalStateId of the creating global state
When two states have a common ancestor, it is safe to share NameRefs
that refer to names in the common ancestor. Before this commit, this
wasn't allowed because NameRefs remembered the id of the immediate state
it dervives from. This led to us triggering enforce in the following
situation when it was in fact safe:

```
auto gs1 = createInitialGlobalState(...);
gs1->enterNameUtf8("demo");
auto gs2 = gs1->deepCopy();
auto gs3 = gs1->deepCopy();
auto nameRef = gs2->enterNameUTF8("demo");
nameRef.data(gs3); // this used to ENFORCE
```