Silence warnings caused by redefining methods.
Suppress the warnings emitted by Ruby when redefining methods by
temporarily setting $VERBOSE to false if it is enabled.

Add a method `T::Configuration.without_ruby_warnings`, which takes a
block and sets $VERBOSE to false, runs the block, and then restores the
original value of $VERBOSE.

This fixes a number of different places in sorbet-runtime that redefine
methods and cause Ruby to emit warnings in verbose mode.