Fix crash with --symbol-table-full (#1288)
* Don't try to take the path of a file that does not exist

Fix #1069.

Synthesized methods like the ones in the sig builder have file id = 0
as their file (through Loc::none). File 0 is a null pointer in the file
table so it's not safe to dereference it.

In debug mode, we were dereferencing it in our FileData ENFORCE, in
release mode, we were crashing later.

The cli phases test that was supposed to catch this crash wasn't doing
its job since the exit status for the test script is not checked. Only
the diff is checked, which crashing did not affect.

* Enforce non null when taking FileData from ref