Allow RBI files to show up in test_corpus
This is kind of janky, but it works well enough for what I need it to do
now. In the normal pipeline, we'd never emit cfg / infer errors within
an RBI file, because we never run typecheckOne on an RBI file. Rather
than make the test harness do this too, I just made it so that my RBI
files don't have an errors in cfg / infer.