Convert partially-sigged test to testdata test (#1302)
* Allow RBI files to show up in test_corpus

This is kind of janky, but it works well enough for what I need it to do
now. In the normal pipeline, we'd never emit cfg / infer errors within
an RBI file, because we never run typecheckOne on an RBI file. Rather
than make the test harness do this too, I just made it so that my RBI
files don't have an errors in cfg / infer.

* Add a test that replaces partially-sigged

This test aims to show that adding a sig in sorbet-typed overrides
having no sig that we get from reflection.

* Remove partially-sigged

* Whoops, removed the wrong test

* Remove TODOs and document what the current state is (#1304)