Redact unique singleton class numbers
These numbers change for every symbol-table snapshot test when an RBI is
added or deleted, which is quite noisy.

I re-used the censorRawLocsWithinPayload option to implement this, which
is kind of wrong because it's censoring unique name numbers now, instead
of raw locs. I'm happy to change the name of the option to account for
this, but if you're fine with the weird name I'm fine not changing it.