Make publish script fail when `file` is missing
This will make publish.sh fail fast when the `file` command isn't
available. As far as I can tell, there's no way to make the bash
pattern:

    some_good_cmd "$(some_bad_command)"

fail fast by enabling a flag. In every case it will need to be
refactored to

    x="$(some_bad_command)"
    some_good_cmd "$x"

which as a Bash enthusiast is deeply saddening. shellcheck doesn't warn
for this, no future version of bash solves this, etc. I don't think it's
worth patently avoiding this, because it can make things more verbose
and coming up with names can be hard, but it's something to be aware of.