Optimize containers of untyped
If we have a container of T.untyped, there's no need to walk the
container when doing typechecking. Optimize by creating
special-purpose subclasses for T.untyped that do no checking.

We opt for special subclasses instead of a generic
`T::Types::UntypedEnumerable` or degrading to a raw `T::Types::Simple`
so that code that inspects types can still inspect the precise
declared type in a compatible way.