Send payload files and files not present editor-side to supported clients with `sorbet:` URIs (#1442)
Internally, Sorbet uses `https://github.com/sorbet/sorbet/..." paths for payload RBIs. Requesting "go-to-definition" on a method described in a payload RBI (e.g., Array.select) results in Sorbet sending back one of these URLs as the URI for the file containing the definition. This works perfectly fine on sorbet.run, but fails in VS Code as VS Code does not support https:// URLs.

In a similar vein, VS Code errors when running "go to definition" on a method that is defined in a file not on the local file system. This is an issue at Stripe, as we have some codegenned files present on the computer running Sorbet that are not present on the laptop running VS Code.

This change causes Sorbet to use sorbet: URIs for the above two categories of files. It does so only if the editor advertises that it understands the URI during initialization.

This change also adds a sorbet/readFile request to Sorbet, which the VS Code extension uses to read the contents of sorbet: files. I initially encoded file data into the URIs, but VS Code froze up when dealing with 700+KB URIs.

Finally, this change add a new option to Sorbet that specifies which folders should be sent as sorbet: URIs to clients that support that URI scheme.