klass -> representedClass
There are actually two classes in play here. The recvType could be a
ClassType because it's actually a `T.class_of(A)` for some A. But
getRepresentedClass is designed to unwrap this type to get a SymbolRef
corresponding to the class inside the `T.class_of` (this is how `===`
works, because the constant literal `A` has type `T.class_of(A)`, so
`A.===(x)` has `T.class_of(A)` for it's recvType).

This is confusing, because I'm about to add a case to `===` when the
recvType **doesn't** have a represented class, namely `false.===(x)`.

Accordingly, I've renamed the variable.