Update knowledge when singleton used in x.===(y) (#1451)
* tp1, tp2 -> argType, recvType

This makes this case more consistent with the other cases in
updateKnowledge.

* klass -> representedClass

There are actually two classes in play here. The recvType could be a
ClassType because it's actually a `T.class_of(A)` for some A. But
getRepresentedClass is designed to unwrap this type to get a SymbolRef
corresponding to the class inside the `T.class_of` (this is how `===`
works, because the constant literal `A` has type `T.class_of(A)`, so
`A.===(x)` has `T.class_of(A)` for it's recvType).

This is confusing, because I'm about to add a case to `===` when the
recvType **doesn't** have a represented class, namely `false.===(x)`.

Accordingly, I've renamed the variable.

* Update knowledge when singleton used in x.===(y)

This PR adds two tests:

- singleton_if_exhaustiveness.rb ALREADY passed on master
- singleton_case_exhaustiveness did not pass on master

I thought it was weird that these two cases diverged from each other, so
I've updated the case for `===` to handle when the reciever is a
singleton value.

This is pre-work for supporting Opus::Enum as a singleton with
exhaustiveness checks.