Emit "missing returns" from decl_builder (#1450)
* Emit "missing returns" from decl_builder

This solves a problem reported on Slack where a custom configuration
handler can't catch this particular error.

The tests don't change, because all BuilderError's are rescued and
handled with the sig_builder_error_handler. The
sig_builder_error_handler used to dispatch to sig_error and from there
raise an ArgumentError. The test was already asserting that there would
be an ArgumentError, so because BuilderError's already always become
ArgumentError from the default handler, the existing test doesn't need
to change.

* Remove unused method

* Fix bug!