Support Singletons in updateKnowledge
This change affects Sorbet's updateKnowledge function: if a class
derives from Singleton (a class in the Ruby standard library[1]) and is
final, then we can know that not being equal to an **instance** of this
type is the same as not being a subtype of that type at all.

[1] https://ruby-doc.org/stdlib-2.4.5/libdoc/singleton/rdoc/Singleton.html

As far as I can tell, support for this class was added in Ruby 2.1.

Much of the pre-work actually landed in a separate PR, to expand support
for updating knowledge of the existing singletons (true, false, nil) in
Ruby's === operator: #1451