Fix crash with sealed
I cargo culted setting singletonClass as sealed from above, but thinking
about it, it doesn't make sense to do so. If a class MyClass is sealed,
it must be the case that on MyClass.singleton_class there is a method
called sealedClassesList.

So if MyClass.singleton_class is also marked sealed, then
MyClass.singleton_class.singleton_class should have a method
sealedClassesList, which there wasn't, and I don't think it ever makes
sense to.

This manifested as a crash in the program before becaise it hit one of
my ENFORCEs somewhere.