Fix string formatting issue
So fmt::Format was actually returning a string_view instead of a string.
This makes sense, until we append content with it. Then, we get weird
output where the the sigString is "overwritten" with the docString.
Definitely something to do with how string_view references and
fmt::format works. I'm not going to investigate further, as I need to
sleep.