Add type bounds to generics (#1392)
* Support type bounds

When first processed, bounds are initially set to `T.untyped`. This
mirrors the previous behavior of setting the `resultType` to
`T.untyped`, and fixing it in a later pass. An alternate approach would
be to use top and bottom for the upper and lower bounds, but this leads
to problems with file ordering.

`test/testdata/todo/fixed_ordering.rb` is a good example of this.

* Move the bounds check on type applications to `typed: false`

* Fix comment in test

* Simplify bounds checking for type_member declarations

* Validate type_member declarations with no explicit bounds

* Format

* Refer to the parent in type_template declarations for enums