Move T::Types::OpusEnum lower in the list (#1570)
Introducing this earlier in the list caused the load order to change in
pay-server (because referencing `Opus::Enum` causes `lib/enum/enum.rb`
to be required), which seems to have revealed a bug in final methods.

The loading bug reproduces with just `puts Opus::Enum` in a file that
mentions `final!`, so I'm working on identifying and fixing the root
cause. This should let us bump sorbet-runtime in pay-server in the mean
time so that people can actually use Opus::Enum in type syntax.