Silence mocha errors
We get errors from the ruby interpreter that say "hey, undefining
initialize is bad, consider not doing that" because under the hood
`expects` is going to undefine and then re-define the stubbed method.

Instead, let's stub `new` so that the ruby VM doesn't complain at us.
This means that `rake test` doesn't print anything to stdout except the
actual minitest output.