Fix segfault
The reason this happens is because we it++ and call line = ... but
do not check if we have reached SOF yet. This is bad, and is my bad
because do while loops arent't as nice for these conditions. I re-used
the while loop, and reverted back to using *it instead of line so that
we don't make weird assignment calls just for readability.