Add test inspired by some pay-server usages
I found some places in pay-server that were essentially doing this but
with manually crafted type aliases and union types instead of sealed.

I wanted to see if sealed would be able to support this and luckily
enough it behaves just the way I expected it to, so I figured I'd add a
test.