Assorted sorbet.org/docs/ changes (#1632)
* docs/final.md: Shorten sidebar label

This makes it fit onto oneline. This was previously our only sidebar
item that didn't fit on one line.

* docs/procs.md: Follow up to changes in #1560

I was thinking about leaving these as feedback on the PR itself, but I
figured it'd be faster for me to just make them in a separate commit.

cc @connorshea

* docs/final.md: Use more words

* sidebars.json: Reorder pages

This is updated to be more in line with the kinds of questions people
get. Also moving tuples and shapes to "Experimental Features" to
discourage their use a little more.

* docs/abstract.md: Use ampersand in sidebar_label

* docs/procs.md: Make sure the word "block" appears in the sidebar

People usually know they're searching for blocks, not procs.

* docs/tstruct.md: Show T::Struct more prominently

* docs/abstract.md: Assorted edits

Most importantly: softens the language about some things are only
implemented in the runtime, as now most of these errors are implemented
statically, with variance checks being the last big holdout.

* docs/override-checking.md: Flesh it out a little

Also invert it to say "mostly statically checked" instead of "mostly
runtime."

* Add some "What's next?" sections to cross link

* Update website/docs/procs.md

Co-Authored-By: Connor Shea <connor.james.shea@gmail.com>

* sidebars.json: Move type-aliases earlier

This is used referenced by both exhaustiveness (and soon to be sealed)
so I'd like to have introduced it first before using it.

* yarn prettier

* typo

* typo

* Add docs for sealed classes (#1634)

* Add docs/sealed.md

This is mostly just a copy / paste from the shipped email, except I
removed the reference Brian Goetz's Data Classes article and replaced it
with an example ADT that I used in a talk about ADTs once.

* Add more links to sealed

* Add "What's next?" section