Add -p cfg-raw (#1728)
* Make cfg-view.sh work on Linux

* Add -p cfg-raw (#1729)

* Add -p cfg-raw

It's pretty rough around the edges, in the sense that it hasn't been
refined and tuned for a variety of CFGs. Even for tiny programs it
generates really big graphs, which graphviz tends to do a poor job of.

But the thing is, at least it's searchable, which means that even if the
layout is terribly hard to follow, it's still useful in the sense that
it gives names to instructions and their fields. So I imagine it might
be best used in combination with (rather than instead of) `-p cfg`.

* Make update_exp_files fail fast, before building Ruby with bazel

* Update test harness