Some more completion.cc refactors (#1770)
* Early exit to reduce nesting

* Use more auto

* Add `return` to get it to format on two lines

* Remove OrType case

It's never the case that a DispatchResult component is an OrType. I
verified this by going into calls.cc where the DispatchComponent
receiver is populated and adding an ENFORCE to verify.

* Move findDocumentation implementation elsewhere

It was defined in completion.cc, but also used in hover.cc. That's a bad
practice in my opinion.