LSP: Move fast path decision + initialGS to preprocessor thread. (#1732)
Moves all initialGS-related logic into a new preprocessor thread. The preprocessor thread passes the fast path decision, changed files, new file hashes, indexed trees, and clones of initialGS in the event of a slow path to the main processor thread.

Updates are eagerly committed to initialGS. To handle the fast path decision in the event of a merge, I've added the ability to move the initialGS "back in time" (by undoing file vector updates and restoring old file hashes) to make a fast path decision in the past. There's one hack: I manually track if a particular update adds a new file to the file vector, as I am unable to safely 'remove' a file. This can be removed when we switch to using tombstones and can get an accurate live file count.

In a subsequent PR, this architecture will enable us to cancel any running slow paths when a new edit comes in that would take the fast path.