Unattribute a comment in resolver.cc
After a year on the team, I never did this, and I don't think we benefit
that much from it.

On second thought, does it even make sense for an abstract method to
have a default argument? It won't apply to the child... Maybe we should
just make this an error?