Remove a whole bunch of TODO(jez) comments (#1838)
* Remove some TODOs that blame to me

After sitting in the codebase for more 6 months, I don't think these are
valuable to ever do. Things work just fine.

* Remove TODOs from namer.cc

There was a time where I thought this was a hack. The more I think about
it, I think it's quite fine.

* Unattribute a comment in resolver.cc

After a year on the team, I never did this, and I don't think we benefit
that much from it.

On second thought, does it even make sense for an abstract method to
have a default argument? It won't apply to the child... Maybe we should
just make this an error?

* Answer two comments in resolver: No and no

* Remove todo to recursively find sorbet/config

At some point we agreed this was too much action at a distance in a
place we wanted to keep simple, and I forgot to remove the TODO