Thread isPrivateOk through all data structures
We already had a concept of PRIVATE_OK on ast::Expression. We were
previously only using it in autogen in one place (arguably a place that
didn't *really* need to be using it) but it does mean that this change
for example doesn't have to touch Desugar.cc.

Note that we expose it all the way through to the lsp::SendReponse so
that we can recall it in completion.cc in a future change.