Factor out doTextDocumentCompletion
I plan to use this helper to add another completion-related assertion
(asserting the result of accepting a single completion item result).

It does some client-specific things like treating the list as being
sorted by sortText or label, and could grow other things that a client
would need to implement to be compliant.