Use textEdit instead of insertText for completion
This gives us control over how to edit the document when a completion
result is taken, rather than having VS Code "figure it out."

VS Code takes some serious liberties with `insertText`:

    /**
     * ...
     *
     * The `insertText` is subject to interpretation by the client side.
     * Some tools might not take the string literally. For example
     * VS Code when code complete is requested in this example `con<cursor position>`
     * and a completion item with an `insertText` of `console` is provided it
     * will only insert `sole`. Therefore it is recommended to use `textEdit` instead
     * since it avoids additional client side interpretation.
     */

In particular, for something like this:

    x.fo
    #   ^ completion: foo

VS Code will delete the `fo` and insert `foo`. But if instead we send
the insertText as just `o` (i.e., remove the `fo` prefix), it will
**still** delete the `fo` and **only** insert `o`.

That's annoying to deal with, because it would mean that to write tests
for snippets, we'd have to re-implement the same heuristic that VS Code
uses to handle `insertText`. Rather than do that, we can just tell VS
Code exactly how to edit the document by sending a `TextEdit` object,
which has a `Range` to replace and a `string` to replace it with.

This will be tested in a stacked PR that adds the ability to test the
state of the document after a completion edit is performed.