Lie and say textDocument/completion is delayable
By our own definition, textDocument/completion is not delayable (need to
finish processing the current edit before servicing it, because the
input references position information).

But the thing is, the common case is that completion requests come in
between two edits that if we could merge them, we could take the fast
path. So for the sake of hacking responsiveness into the IDE, we're
compromising on our principles and pretending that
textDocument/completion is delayable.